Added Life and Level decrease/increase events.

+++ src/EventHorizon.Plugin.Zone.System.Combat/Events/Level/RunLevelUpForEntityEvent.cs
@@ -5,7 +5,7 @@ namespace EventHorizon.Plugin.Zone.System.Combat.Events.Level
 {
     public struct RunLevelUpForEntityEvent : INotification
     {
-        public int EntityId { get; set; }
+        public long EntityId { get; set; }
         public LevelProperty Property { get; set; }
     }
 }
